Q2/backend/rag/requirements.txt
# Core RAG libraries
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.12
openai==1.10.0

# Vector stores
chromadb==0.4.22
faiss-cpu==1.7.4

# API framework
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Environment management
python-dotenv==1.0.0

# Document processing
pypdf==4.0.1
python-docx==1.1.0
beautifulsoup4==4.12.3
requests==2.31.0

# Evaluation
ragas==0.1.1
datasets==2.16.1

# Data processing
numpy==1.26.3
pandas==2.2.0
scikit-learn==1.4.0

# Embeddings
sentence-transformers==2.2.2

# Utilities
tqdm==4.66.1
asyncio==3.4.3
aiofiles==23.2.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4

# Logging
loguru==0.7.2

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Metrics and monitoring
prometheus-client==0.19.0
